add: Add task startup conditions

+++ lib/tasks/task_sample.rake
@@ -1,6 +1,8 @@
 desc 'Automated code analysis and test execution'
 task :rspec_test do
-  sh 'bundle exec rubocop'
-  sh 'bundle exec rspec'
-  sh 'rm -rf public/uploads/tmp'
+  if Rails.env.development?
+    sh 'bundle exec rubocop'
+    sh 'bundle exec rspec'
+    sh 'rm -rf public/uploads/tmp'
+  end
 end
